+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - domains and domains_url search index routing were added for german locale
     - CategoryRepositoryTest was fixed for third domain
     - translations were copied from en language and prefixed with (de)
+- [#81 - added google tag manager scripts](https://github.com/shopsys/demoshop/pull/81)
 
 ### Changed
 - [#1 - Basic changes in docs, readme etc. after copying from project-base](https://github.com/shopsys/demoshop/pull/1) : [@LukasHeinz]